+++ app/settings/tokens/-components/token-form/component.ts
@@ -12,7 +12,6 @@ export default class TokenForm extends Component {
     // Optional arguments
     token?: Token; // If not provided, new token created by validated-model-form
     disabled?: boolean;
-    analyticsScope?: string;
 
     // Private properties
     @service store!: DS.Store;