+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `getHref` - get an href from a `Link`
     - `getRelatedHref` - get an href from a `Relationship`
     - `tuple` - create a strictly-typed [tuple](https://www.typescriptlang.org/docs/handbook/basic-types.html#tuple)
+
 ### Changed
 - Models:
     - `osf-model` - use proper types for `links` and `relationshipLinks` attributes
@@ -21,12 +22,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `developer-app` - extend `links` types to include links specific to developer apps
     - `collection` - improve types for choices fields
     - `collected-metadatum` - improve types for choice fields
-- Components:
     - `user-setting` - Added `requestExport()` function, `deactivationRequested` field, and `export` link
+- Components
+    - `osf-link` - used to be `link`
+        - `@onClick` parameter used to be `@onclick`
     - `<ContributorList>` - modified to take in different parameters
         - `node` (required)
         - `shouldTruncate` (default true)
         - `shouldLinkUsers` (default false)
+- Tests
+    - Using new `click` handler everywhere in main app to verify `data-analytics-name` usage
+
+### Removed
+- Components:
+    - `global-link-to`
 
 ## [19.1.0] - 2019-01-23
 ### Added